refactor: factor out figures utils
